+++ app/src/main/java/com/vanpra/composematerialdialogdemos/demos/BasicDialog.kt
@@ -3,6 +3,7 @@ package com.vanpra.composematerialdialogdemos.demos
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.LocationOn
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.stringResource
 import com.vanpra.composematerialdialogdemos.DialogAndShowButton
 import com.vanpra.composematerialdialogdemos.R
 
@@ -43,6 +44,15 @@ fun BasicDialogDemo() {
         buttons {
             negativeButton("No Thanks")
             positiveButton("Turn On Speed Boost")
+        }
+    }
+
+    DialogAndShowButton(buttonText = "Basic Dialog With input") {
+        title(res = R.string.input_dialog_title)
+        input(label = "Name", hint = "Jon Smith")
+        buttons {
+            negativeButton("Cancel")
+            positiveButton("Ok")
         }
     }
 }