+++ core/src/main/java/com/vanpra/composematerialdialogs/MaterialDialog.kt
@@ -4,14 +4,28 @@ import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.text.BasicText
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.preferredHeight
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.foundation.layout.wrapContentWidth
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.material.TextField
-import androidx.compose.runtime.*
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.runtime.savedinstancestate.savedInstanceState
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -21,6 +35,7 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.layout.Layout
 import androidx.compose.ui.layout.layoutId
 import androidx.compose.ui.platform.AmbientDensity
+import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
@@ -271,10 +286,11 @@ class MaterialDialog(private val autoDismiss: Boolean = true) {
         val index by mutableStateOf(positiveEnabled.size)
         var valid by mutableStateOf(allowEmpty)
 
-        remember { positiveEnabled.add(index, allowEmpty) }
-
-        if (waitForPositiveButton) {
-            callbacks.add { onInput(text) }
+        remember {
+            positiveEnabled.add(index, allowEmpty)
+            if (waitForPositiveButton) {
+                callbacks.add { onInput(text) }
+            }
         }
 
         Column(modifier = Modifier.padding(start = 24.dp, end = 24.dp, top = 8.dp, bottom = 8.dp)) {
@@ -285,7 +301,6 @@ class MaterialDialog(private val autoDismiss: Boolean = true) {
                     if (!waitForPositiveButton) {
                         onInput(text)
                     }
-
                     // Have to make temp list in order for state to register change
                     val tempList = positiveEnabled.toMutableList()
                     valid = if (text == "" && allowEmpty) {
@@ -298,12 +313,13 @@ class MaterialDialog(private val autoDismiss: Boolean = true) {
                     tempList[index] = valid
                     positiveEnabled = tempList
                 },
-                label = { BasicText(label) },
+                label = { Text(label, color = MaterialTheme.colors.onBackground.copy(0.8f)) },
                 modifier = Modifier.fillMaxWidth(),
-                placeholder = { BasicText(hint) },
+                placeholder = { Text(hint, color = MaterialTheme.colors.onBackground.copy(0.5f)) },
                 isErrorValue = !valid,
                 visualTransformation = visualTransformation,
                 keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Text),
+                textStyle = TextStyle(MaterialTheme.colors.onBackground, fontSize = 16.sp)
             )
 
             if (!valid) {